fix: Move ts rest config
