+++ src/features/auth/LoginForm.tsx
@@ -7,11 +7,11 @@ import { ClientInferRequest } from '@ts-rest/core';
 import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 
+import { Contract } from '@/api/contract';
 import { FieldInput } from '@/components/FieldInput';
 import { useToastError } from '@/components/Toast';
 import { useLogin } from '@/features/auth/service';
 import { DemoLoginHint } from '@/features/demo-mode/DemoLoginHint';
-import { Contract } from '@/lib/tsRest/contract';
 
 type LoginFormProps = BoxProps & {
   onSuccess?: () => void;
